+++ utils/test-utils/classes/TestClassRunnerTestClass.ts
@@ -1,17 +1,40 @@
 export class TestClassRunnerTestClass {
 
-  public beforeAll():void {}
+  public static invokator:any = null;
 
-  public static beforeClass():void {}
+  public beforeAll():void {
+    TestClassRunnerTestClass.invokator.notify("beforeAll");
+  }
 
-  public before():void {}
+  public static beforeClass():void {
+    TestClassRunnerTestClass.invokator.notify("beforeClass");
+  }
 
-  public test():void {}
+  public before():void {
+    TestClassRunnerTestClass.invokator.notify("before");
+  }
+
+  public test():void {
+    TestClassRunnerTestClass.invokator.notify("test");
+  }
   
-  public afterAll():void {}
+  public test2():void {
+    TestClassRunnerTestClass.invokator.notify("test2");
+  }
 
-  public static afterClass():void {}
+  public test3():void {
+    TestClassRunnerTestClass.invokator.notify("test3");
+  }
 
-  public after():void {}
+  public afterAll():void {
+    TestClassRunnerTestClass.invokator.notify("afterAll");
+  }
 
+  public static afterClass():void {
+    TestClassRunnerTestClass.invokator.notify("afterClass");
+  }
+
+  public after():void {
+    TestClassRunnerTestClass.invokator.notify("after");
+  }
 }